+++ src/slave(Pi-Pico)/src/global_vars.c
@@ -4,16 +4,16 @@ const struct __MOTORs_PIN
 {
     struct __MOTOR_PIN
     {
-        unsigned short int terminal_1;
-        unsigned short int terminal_2;
+        uint8_t terminal_1;
+        uint8_t terminal_2;
     } motor_right, motor_left;
 } MOTORs_PINs = {16, 17, 18, 19};
 
-const short int MOTORs_PINs_motor_right_terminal_1 = MOTORs_PINs.motor_right.terminal_1;
-const short int MOTORs_PINs_motor_right_terminal_2 = MOTORs_PINs.motor_right.terminal_2;
-const short int MOTORs_PINs_motor_left_terminal_1 = MOTORs_PINs.motor_left.terminal_1;
-const short int MOTORs_PINs_motor_left_terminal_2 = MOTORs_PINs.motor_left.terminal_2;
+const uint8_t MOTORs_PINs_motor_right_terminal_1 = MOTORs_PINs.motor_right.terminal_1;
+const uint8_t MOTORs_PINs_motor_right_terminal_2 = MOTORs_PINs.motor_right.terminal_2;
+const uint8_t MOTORs_PINs_motor_left_terminal_1 = MOTORs_PINs.motor_left.terminal_1;
+const uint8_t MOTORs_PINs_motor_left_terminal_2 = MOTORs_PINs.motor_left.terminal_2;
 
 const bool GPIO_ON = 1;
 const bool GPIO_OFF = 0;
-const unsigned short int GREEN_LED = 25;
+const uint8_t GREEN_LED = 25;
